fix: DTO 디렉토리 구조 수정, 필요 없는 내부 메서드 정리
