+++ src/main/java/com/sprint/mission/discodeit/dto/BinaryContentDto.java
@@ -1,0 +1,12 @@
+package com.sprint.mission.discodeit.dto;
+
+import java.util.UUID;
+
+public record BinaryContentDto(
+    UUID id,
+    String fileName,
+    Long size,
+    String contentType
+) {
+
+}
