+++ src/main/java/com/sprint/mission/discodeit/dto/ChannelDto.java
@@ -1,0 +1,17 @@
+package com.sprint.mission.discodeit.dto;
+
+import com.sprint.mission.discodeit.domain.ChannelType;
+import java.time.Instant;
+import java.util.List;
+import java.util.UUID;
+
+public record ChannelDto(
+    UUID id,
+    ChannelType type,
+    String name,
+    String description,
+    List<UserDto> participants,
+    Instant lastMessageCreatedAt
+) {
+
+}
