+++ src/main/java/com/sprint/mission/discodeit/dto/MessageDto.java
@@ -1,0 +1,17 @@
+package com.sprint.mission.discodeit.dto;
+
+import java.time.Instant;
+import java.util.List;
+import java.util.UUID;
+
+public record MessageDto(
+    UUID id,
+    Instant createdAt,
+    Instant updatedAt,
+    String content,
+    UUID channelId,
+    UserDto author,
+    List<BinaryContentDto> attachments
+) {
+
+}
