+++ src/main/java/com/sprint/mission/discodeit/dto/ReadStatusDto.java
@@ -1,0 +1,13 @@
+package com.sprint.mission.discodeit.dto;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record ReadStatusDto(
+    UUID id,
+    UUID userId,
+    UUID channelId,
+    Instant lastReadAt
+) {
+
+}
