+++ src/main/java/com/sprint/mission/discodeit/dto/UserDto.java
@@ -1,0 +1,13 @@
+package com.sprint.mission.discodeit.dto;
+
+import java.util.UUID;
+
+public record UserDto(
+    UUID id,
+    String username,
+    String email,
+    BinaryContentDto profile,
+    boolean online
+) {
+
+}
