+++ src/main/java/com/sprint/mission/discodeit/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.request;
 
 public record LoginRequest(
     String username,
