+++ src/main/java/com/sprint/mission/discodeit/dto/request/MessageCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.request;
 
 import java.util.UUID;
