+++ src/main/java/com/sprint/mission/discodeit/dto/request/PublicChannelUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.request;
 
 public record PublicChannelUpdateRequest(
     String newName,
