+++ src/main/java/com/sprint/mission/discodeit/dto/request/ReadStatusCreateRequest.java
@@ -1,0 +1,12 @@
+package com.sprint.mission.discodeit.dto.request;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record ReadStatusCreateRequest(
+    UUID userId,
+    UUID channelId,
+    Instant lastReadAt
+) {
+
+}
