+++ src/main/java/com/sprint/mission/discodeit/dto/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.request;
 
 public record UserCreateRequest(
     String email,
