+++ src/main/java/com/sprint/mission/discodeit/dto/request/UserStatusCreateRequest.java
@@ -1,0 +1,11 @@
+package com.sprint.mission.discodeit.dto.request;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record UserStatusCreateRequest(
+    UUID userId,
+    Instant lastActiveAt
+) {
+
+}
