+++ src/main/java/com/sprint/mission/discodeit/dto/request/UserUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.request;
 
 public record UserUpdateRequest(
     String newUsername,
